Rewis3d_Reconstruction/pyproject.toml
[project]
name = "map-anything"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "huggingface_hub",
    "hydra-core",
    "natsort",
    "opencv-python-headless==4.10.0.84",
    "orjson",
    "pillow-heif",
    "plyfile",
    "python-box",
    "safetensors",
    "tensorboard",
    "tqdm",
    "torchaudio", ## Temp for AWS
    "trimesh",
    "uniception",
    "open3d>=0.19.0",
    "gradio>=5.47.0",
    "spaces>=0.41.1",
    "opencv-python>=4.12.0.88",
    "moviepy<2.0",
    "pycolmap>=3.13.0",
    "evo>=1.34.0",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
